% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{download_all_data}
\alias{download_all_data}
\title{Download all datasets}
\usage{
download_all_data(save_path)
}
\arguments{
\item{save_path}{String, path to save the data}
}
\description{
Download all datasets
}
