% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{load_data}
\alias{load_data}
\title{Load clinical trial data}
\usage{
load_data(task, phase)
}
\arguments{
\item{task}{String, dataset name, such as 'mortality_rate'}

\item{phase}{String, clinical trial phase, such as 'Phase 1', 'Phase 2', 'Phase 3', 'Phase 4'}
}
\value{
List, containing train_df, valid_df, test_df, num_classes and tabular_input_dim
}
\description{
Load clinical trial data for specified task and phase by calling Python's trialbench.load_data function
}
